+++ plone/app/event/recurrence.py
@@ -14,7 +14,6 @@ from zope.component import adapts
 from zope.interface import Interface
 from zope.interface import implements
 from zope.publisher.interfaces.browser import IBrowserRequest
-import itertools
 
 try:
     from repoze.zope2.publishtraverse import DefaultPublishTraverse
@@ -51,8 +50,18 @@ class RecurrenceSupport(object):
         """Return all occurrences of an event, possibly within a start and end
         limit.
 
+        :param range_start: Optional start datetime, from which you want
+                            occurrences be returned.
+        :type range_start: Python datetime
+        :param range_end: Optional start datetime, from which you want
+                          occurrences be returned.
+        :type range_end: Python datetime
+        :returns: List of occurrences, including the start event.
+        :rtype: IEvent or IOccurrence based objects
+
         Please note: Events beginning before range_start but ending afterwards
                      won't be found.
+        TODO: really?
 
         TODO: test with event start = 21st feb, event end = start+36h,
         recurring for 10 days, range_start = 1st mar, range_end = last Mark
@@ -85,11 +94,8 @@ class RecurrenceSupport(object):
                 start=start,
                 end=start + duration).__of__(self.context)
 
-        # TODO: check performuance.
-        # Here, the generator is converted to a list.
-        events = map(get_obj, starts)
-
-        return events
+        for start in starts:
+            yield get_obj(start)
 
 
 class OccurrenceTraverser(DefaultPublishTraverse):
@@ -104,8 +110,8 @@ class OccurrenceTraverser(DefaultPublishTraverse):
         context = self.context
         dateobj = guess_date_from(name, context)
         if dateobj:
-            occurrence = IRecurrenceSupport(context).occurrences(
-                range_start=dateobj)[0]
+            occs = IRecurrenceSupport(context).occurrences(range_start=dateobj)
+            occurrence = occs.next()
             occ_acc = IEventAccessor(occurrence)
             if is_same_day(dateobj, occ_acc.start):
                 return occurrence