+++ plone/app/event/tests/test_icalendar.py
@@ -33,6 +33,7 @@ class ICalendarExportTestDX(AbstractSampleDataEvents):
     layer = PAEventDX_INTEGRATION_TESTING
 
     def event_factory(self):
+        DXEventAccessor.portal_type = 'plone.app.event.dx.event'
         return DXEventAccessor.create
 
     def traverser(self, context, request):
@@ -60,8 +61,6 @@ class ICalendarExportTestDX(AbstractSampleDataEvents):
             'BEGIN:VCALENDAR',
             'VERSION:2.0',
             'PRODID:-//Plone.org//NONSGML plone.app.event//EN',
-            'X-WR-CALNAME:Now Event',  # calendar name == event title
-            'X-WR-RELCALID:',
             'X-WR-TIMEZONE:Europe/Vienna',
             'BEGIN:VEVENT',
             'SUMMARY:Now Event',
@@ -126,7 +125,6 @@ class ICalendarExportTestDX(AbstractSampleDataEvents):
             'BEGIN:VCALENDAR',
             'VERSION:2.0',
             'PRODID:-//Plone.org//NONSGML plone.app.event//EN',
-            'X-WR-CALNAME:Plone site',  # calendar name == plone site title
             'X-WR-TIMEZONE:Europe/Vienna',
             # whole_day event
             'BEGIN:VEVENT',